requirements.txt
# Digital FTE - Personal AI Employee
# Python Dependencies

# Google API for Gmail Watcher
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0

# File watching
watchdog>=3.0.0

# Browser automation for WhatsApp
playwright>=1.40.0

# Environment management
python-dotenv>=1.0.0

# HTTP requests
requests>=2.31.0

# Date/time utilities
python-dateutil>=2.8.2

# Process management (optional)
psutil>=5.9.0

# MCP Server
fastmcp>=0.1.0
jinja2>=3.1.0

# Push Notifications (Mobile)
pywebpush>=1.14.0
py-vapid>=1.9.0
cryptography>=41.0.0

# Firebase Cloud Messaging (FCM) for Mobile Push
firebase-admin>=6.0.0

# Telegram Bot (fallback notifications)
python-telegram-bot>=20.0

# FastAPI (for notification API)
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Markdown processing (for LinkedIn scheduler)
python-frontmatter>=1.0.0
linkedin-api>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Spotify integration (optional)
spotipy>=2.23.0
